pyproject.toml
[tool.poetry]
name = "filmshowio"
version = "0.0.1"
description = "A virtual cinema hall"
authors = ["Abhinav Kumar <abhinav.vikash2@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
djoser = "^2.1.0"
django-cors-headers = "^3.14.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
